+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `getSms` to `SmsClient` for searching for a single message by id.
 - Added `ConversionClient` and the ability to interact with the Nexmo Conversion API.
 
+### Changed
+- `TalkNcco` now uses `VoiceName` internally instead of `String` for the `voiceName` property.
+
 ## [3.5.0] - 2018-05-29
 ### Changed
 - Updated `VerifyClient` to use the JSON endpoints instead of XML.